pyproject.toml
[project]
name = "mongodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.5",
    "dotenv>=0.9.9",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pymongo>=4.15.4",
    "seaborn>=0.13.2",
    "streamlit-avatar>=0.1.3",
]
